requirements.txt
annotated-types==0.7.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comtypes==1.4.8
confection==0.1.5
cymem==2.0.8
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
langcodes==3.4.1
language_data==1.2.0
language_tool_python==2.8.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.4.2
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
preshed==3.0.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pypiwin32==223
pyspellchecker==0.8.1
pytesseract==0.3.13
pyttsx3==2.98
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
setuptools==75.5.0
shellingham==1.5.4
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.1
thinc==8.3.2
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
typer==0.13.0
typing_extensions==4.12.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wheel==0.45.0
wrapt==1.16.0
